Extract reporting API module
