+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -113,3 +113,4 @@
 | LWB-R106 | PM operating-quality API module extraction | Hardened | `src/features/workbench/api.ts` still owned the full PM operating-quality Gateway route family plus shared fetch, mutation, URL, and observability helpers. That kept a fast-growing BFF API barrel above 90KB and made PM-quality route changes harder to review without scanning unrelated portfolio, performance, risk, reporting, proof-pack, and construction helpers. | Added `api-client.ts` for shared Workbench Gateway fetch, mutation, URL, error, and bounded observability helpers. Added `pm-operating-quality-api.ts` for policy, score-run, support-summary, fairness-analysis, and review-action Gateway helpers, with `api.ts` re-exporting the same public API for compatibility. Kept all PM-quality request bodies Gateway-shaped and moved the duplicated fairness-analysis body construction into one helper without adding browser-side score, fairness, ranking, HR, client, order, OMS, or execution logic. | Continue splitting the Workbench API layer by coherent Gateway route family only when it reduces review blast radius. Preserve the public `api.ts` barrel until downstream imports are migrated deliberately. | `src/features/workbench/api-client.ts`, `src/features/workbench/pm-operating-quality-api.ts`, `src/features/workbench/api.ts`; validation via Workbench API tests, PM-quality hook tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
 | LWB-R107 | Proof-pack API module extraction | Hardened | After the shared API-client and PM-quality split, `src/features/workbench/api.ts` still owned the proof-pack generation, detail, Markdown, report-input, AI-evidence, and PM-memo Gateway helpers even though proof-pack UI orchestration already lives in `use-proof-pack-actions.ts`. That kept proof-pack mutation/request payloads coupled to unrelated Workbench API families. | Added `proof-pack-api.ts` for proof-pack Gateway helpers and caller headers, with `api.ts` re-exporting the same public functions for compatibility. Kept DPM wave proof-pack posture with the DPM wave surface, and preserved Gateway-shaped proof-pack payloads, idempotency, and bounded AI memo request outputs without adding browser-side proof-pack construction, hash generation, Markdown synthesis, prompt construction, approval, client communication, order, OMS, fills, settlement, or execution behavior. | Continue splitting `api.ts` by route family only where it removes real review coupling. Migrate direct feature imports away from the compatibility barrel in a later mechanical slice if it can be done without expanding behavior. | `src/features/workbench/proof-pack-api.ts`, `src/features/workbench/api.ts`; validation via Workbench API proof-pack tests, proof-pack hook tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
 | LWB-R108 | Outcome-review API module extraction | Hardened | `src/features/workbench/api.ts` still owned outcome-review list, report-input, AI-evidence, advisor narrative, and report-job helpers after proof-pack and PM-quality route families were split out. That kept outcome-review handoff payloads and reporting caller headers coupled to unrelated Workbench API helpers. | Added `outcome-review-api.ts` for outcome-review Gateway helpers and moved shared reporting caller headers into `reporting-api-headers.ts`, with `api.ts` re-exporting the same public functions for compatibility. Preserved Gateway-shaped report-input, AI-evidence, narrative, and report-job payloads without adding browser-side outcome truth, report construction, client communication, approval, order, OMS, fills, settlement, or execution behavior. | Continue splitting `api.ts` by coherent route family only where it removes real review coupling. Preserve the public `api.ts` barrel until downstream imports are migrated deliberately. | `src/features/workbench/outcome-review-api.ts`, `src/features/workbench/reporting-api-headers.ts`, `src/features/workbench/api.ts`; validation via Workbench API outcome-review tests, outcome-review handoff hook tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
+| LWB-R109 | Reporting API module extraction | Hardened | `src/features/workbench/api.ts` still owned report-batch materialization, status, run-once, archived-document metadata, and document-download URL helpers after the reporting operations hook had already been extracted. That kept report/archive request shaping coupled to unrelated Workbench API families and made Gateway reporting changes noisier to review. | Added `reporting-api.ts` for report-batch and archived-document Gateway helpers, reusing `reporting-api-headers.ts`, with `api.ts` re-exporting the same public functions for compatibility. Preserved explicit-portfolio report-batch payloads, source-owned status/run-once posture, archive metadata retrieval, idempotency, and observability labels without adding browser-side batch membership, report generation truth, archive truth, retention decisions, delivery, communication, OMS, or execution behavior. | Continue splitting `api.ts` by coherent route family only where it removes real review coupling. Preserve the public `api.ts` barrel until downstream imports are migrated deliberately. | `src/features/workbench/reporting-api.ts`, `src/features/workbench/reporting-api-headers.ts`, `src/features/workbench/api.ts`; validation via Workbench API reporting tests, report-batch operations hook tests, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
